Restructure packaging for eps
COAND-497
