+++ eps/src/main/java/com/adyen/checkout/eps/EPSComponent.kt
@@ -14,6 +14,7 @@ import com.adyen.checkout.components.base.ComponentEventHandler
 import com.adyen.checkout.components.model.payments.request.EPSPaymentMethod
 import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.eps.EPSComponent.Companion.PROVIDER
+import com.adyen.checkout.eps.internal.provider.EPSComponentProvider
 import com.adyen.checkout.issuerlist.IssuerListComponent
 import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
